Step learning rate scheduler AFTER optimizer
Prior to PyTorch 1.1.0, the learning rate scheduler was expected to be
called before the optimizer’s update; 1.1.0 changed this behavior in a
BC-breaking way. If you use the learning rate scheduler (calling
scheduler.step()) before the optimizer’s update (calling
optimizer.step()), this will skip the first value of the learning rate
schedule. If you are unable to reproduce results after upgrading to
PyTorch 1.1.0, please check if you are calling scheduler.step() at the
wrong time.